Move temp dir set up to the root release file
